# ForkHierarchy/Client/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ForkHierarchy.Server",
    "version": "1.0"
  },
  "paths": {
    "/api/GitHubRepositories/{id}": {
      "get": {
        "tags": [
          "GitHubRepositories"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubRepository"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubRepository"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubRepository"
                }
              }
            }
          }
        }
      }
    },
    "/api/GitHubRepositories": {
      "get": {
        "tags": [
          "GitHubRepositories"
        ],
        "parameters": [
          {
            "name": "fullName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubRepository"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubRepository"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubRepository"
                }
              }
            }
          }
        }
      }
    },
    "/api/QueuedRepositories": {
      "get": {
        "tags": [
          "QueuedRepositories"
        ],
        "parameters": [
          {
            "name": "fullName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedRepository"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedRepository"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedRepository"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "QueuedRepositories"
        ],
        "parameters": [
          {
            "name": "repo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccountType": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "GitHubRepository": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "ghId": {
            "type": "integer",
            "format": "int64"
          },
          "htmlUrl": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "stars": {
            "type": "integer",
            "format": "int32"
          },
          "isFork": {
            "type": "boolean"
          },
          "forksCount": {
            "type": "integer",
            "format": "int32"
          },
          "lastCommit": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "owner": {
            "$ref": "#/components/schemas/GitHubUser"
          },
          "parentId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sourceId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GitHubRepository"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GitHubUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "login": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/AccountType"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "htmlUrl": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "QueuedRepository": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "owner": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "addedAt": {
            "type": "string",
            "format": "date-time"
          },
          "eta": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}